requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pytz>=2021.1
requests>=2.25.1
python-dotenv>=1.0.0
python-pptx>=1.0.2
